refactor : css 위치 수정
