+++ src/component/missionTooltip.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../missionTooltip.css";
+import "../css/roadmap/missionTooltip.css";
 
 function MissionTooltip({ mission, position }) {
   // 툴팁 위치를 커서 옆으로 살짝 이동시키기 위한 스타일
